Created orders module

+++ src/products/products.controller.ts
@@ -1,7 +1,18 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Put,
+  NotFoundException,
+  ParseUUIDPipe,
+} from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { Product } from 'src/db';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -12,17 +23,36 @@ export class ProductsController {
   }
 
   @Get('/:id')
-  public getById(@Param('id') id: string): Product {
-    return this.productsService.getById(id);
+  public getById(@Param('id', new ParseUUIDPipe()) id: string): Product | null {
+    const prod = this.productsService.getById(id);
+    if (!prod) throw new NotFoundException('Product not found');
+    return prod;
   }
 
   @Delete('/:id')
-  public deleteProduct(@Param('id') id: string) {
-    return this.productsService.deleteProduct(id);
+  public deleteProduct(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
+    this.productsService.deleteProduct(id);
+    return { success: true };
   }
 
   @Post('/')
   public createProduct(@Body() productData: CreateProductDTO) {
     return this.productsService.createProduct(productData);
+  }
+
+  @Put('/:id')
+  public editProduct(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ): { success: boolean } {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
+
+    this.productsService.editProduct(id, productData);
+    return { success: true };
   }
 }
